catalog-backend: refactor entityProviders connect
